Create python file for database connection, modified data and Predict page
